requirements.txt
tensorflow
opencv-python
gradio
numpy
fastapi
uvicorn
pillow